requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo;python_version<"3.9"
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
dateparser==1.1.1
Flask==2.2.2
frozenlist==1.3.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.23.2
pandas==1.4.3
pymongo==3.12.3
pyTelegramBotAPI==4.6.1
python-binance==1.0.16
python-dateutil==2.8.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.2
requests==2.28.1
six==1.16.0
timeloop==1.0.2
tzdata==2022.2
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.11
websockets==10.3
Werkzeug==2.2.2
XlsxWriter==3.0.3
yarl==1.8.1
zipp==3.8.1

binance~=0.3